Clip text bounds to effective viewport.
Required moving effective viewport calculation outside of `LayoutManager` into an extension method.
